+++ src/main/java/com/owlmaddie/chat/ChatDataManager.java
@@ -189,7 +189,7 @@ public class ChatDataManager {
             contextData.put("world_moon_phase", moonPhaseDescription);
 
             // Get Entity details
-            MobEntity entity = ServerEntityFinder.getEntityByUUID(player.getServerWorld(), UUID.fromString(entityId));
+            MobEntity entity = (MobEntity)ServerEntityFinder.getEntityByUUID(player.getServerWorld(), UUID.fromString(entityId));
             if (entity.getCustomName() == null) {
                 contextData.put("entity_name", "");
             } else {
@@ -256,7 +256,7 @@ public class ChatDataManager {
                     ParsedMessage result = MessageParser.parseMessage(output_message.replace("\n", " "));
 
                     // Apply behaviors (if any)
-                    MobEntity entity = ServerEntityFinder.getEntityByUUID(player.getServerWorld(), UUID.fromString(entityId));
+                    MobEntity entity = (MobEntity)ServerEntityFinder.getEntityByUUID(player.getServerWorld(), UUID.fromString(entityId));
                     for (Behavior behavior : result.getBehaviors()) {
                         LOGGER.info("Behavior: " + behavior.getName() + (behavior.getArgument() != null ?
                                     ", Argument: " + behavior.getArgument() : ""));